reuse existing `Arbitrary`, `Gen`, and `Cogen` instances
