+++ core/common/src/test/scala/org/typelevel/otel4s/scalacheck/Cogens.scala
@@ -14,77 +14,16 @@
  * limitations under the License.
  */
 
-package org.typelevel.otel4s
+package org.typelevel.otel4s.scalacheck
 
-import org.scalacheck.Arbitrary
 import org.scalacheck.Cogen
-import org.scalacheck.Gen
-import org.scalacheck.Gen.listOf
-import org.scalacheck.Gen.nonEmptyListOf
 import org.scalacheck.rng.Seed
+import org.typelevel.otel4s.Attribute
+import org.typelevel.otel4s.AttributeKey
+import org.typelevel.otel4s.AttributeType
+import org.typelevel.otel4s.Attributes
 
-object arbitrary extends ArbitraryInstances
-trait ArbitraryInstances {
-
-  val nonEmptyString: Gen[String] =
-    Arbitrary.arbitrary[String].suchThat(_.nonEmpty)
-
-  implicit val booleanAttribute: Gen[Attribute[Boolean]] = for {
-    b <- Arbitrary.arbitrary[Boolean]
-    k <- nonEmptyString
-  } yield Attribute(k, b)
-
-  implicit val stringAttribute: Gen[Attribute[String]] = for {
-    s <- nonEmptyString
-    k <- nonEmptyString
-  } yield Attribute(k, s)
-
-  implicit val longAttribute: Gen[Attribute[Long]] = for {
-    l <- Arbitrary.arbitrary[Long]
-    k <- nonEmptyString
-  } yield Attribute(k, l)
-
-  implicit val doubleAttribute: Gen[Attribute[Double]] = for {
-    d <- Arbitrary.arbitrary[Double]
-    k <- nonEmptyString
-  } yield Attribute(k, d)
-
-  implicit val stringListAttribute: Gen[Attribute[List[String]]] = for {
-    l <- nonEmptyListOf(nonEmptyString)
-    k <- nonEmptyString
-  } yield Attribute(k, l)
-
-  implicit val booleanListAttribute: Gen[Attribute[List[Boolean]]] = for {
-    l <- nonEmptyListOf(Arbitrary.arbitrary[Boolean])
-    k <- nonEmptyString
-  } yield Attribute(k, l)
-
-  implicit val longListAttribute: Gen[Attribute[List[Long]]] = for {
-    l <- nonEmptyListOf(Arbitrary.arbitrary[Long])
-    k <- nonEmptyString
-  } yield Attribute(k, l)
-
-  implicit val doubleListAttribute: Gen[Attribute[List[Double]]] = for {
-    l <- nonEmptyListOf(Arbitrary.arbitrary[Double])
-    k <- nonEmptyString
-  } yield Attribute(k, l)
-
-  implicit val attribute: Arbitrary[Attribute[_]] = Arbitrary(
-    Gen.oneOf(
-      booleanAttribute,
-      stringAttribute,
-      longAttribute,
-      doubleAttribute,
-      stringListAttribute,
-      booleanListAttribute,
-      longListAttribute,
-      doubleListAttribute
-    )
-  )
-
-  implicit val attributes: Arbitrary[Attributes] = Arbitrary(
-    listOf(attribute.arbitrary).map(Attributes(_: _*))
-  )
+trait Cogens {
 
   implicit val attributeTypeCogen: Cogen[AttributeType[_]] =
     Cogen[String].contramap(_.toString)
@@ -121,4 +60,7 @@ trait ArbitraryInstances {
 
   implicit val attributesCogen: Cogen[Attributes] =
     Cogen[List[Attribute[_]]].contramap(_.toList)
+
 }
+
+object Cogens extends Cogens
