+++ core/trace/src/test/scala/org/typelevel/otel4s/trace/SpanKindSuite.scala
@@ -19,32 +19,17 @@ package org.typelevel.otel4s.trace
 import cats.Show
 import cats.kernel.laws.discipline.HashTests
 import munit._
-import org.scalacheck.Arbitrary
-import org.scalacheck.Cogen
-import org.scalacheck.Gen
 import org.scalacheck.Prop
+import org.typelevel.otel4s.trace.scalacheck.Arbitraries._
+import org.typelevel.otel4s.trace.scalacheck.Cogens._
+import org.typelevel.otel4s.trace.scalacheck.Gens
 
 class SpanKindSuite extends DisciplineSuite {
-
-  private val spanKindGen: Gen[SpanKind] =
-    Gen.oneOf(
-      SpanKind.Internal,
-      SpanKind.Server,
-      SpanKind.Client,
-      SpanKind.Producer,
-      SpanKind.Consumer
-    )
-
-  private implicit val spanKindArbitrary: Arbitrary[SpanKind] =
-    Arbitrary(spanKindGen)
-
-  private implicit val spanKindCogen: Cogen[SpanKind] =
-    Cogen[String].contramap(_.toString)
 
   checkAll("SpanKind.HashLaws", HashTests[SpanKind].hash)
 
   property("Show[SpanKind]") {
-    Prop.forAll(spanKindGen) { spanKind =>
+    Prop.forAll(Gens.spanKind) { spanKind =>
       val expected = spanKind match {
         case SpanKind.Internal => "Internal"
         case SpanKind.Server   => "Server"
