+++ sdk/trace/src/test/scala/org/typelevel/otel4s/sdk/trace/SdkSpanBackendSuite.scala
@@ -26,7 +26,6 @@ import cats.syntax.traverse._
 import munit.CatsEffectSuite
 import munit.ScalaCheckEffectSuite
 import munit.internal.PlatformCompat
-import org.scalacheck.Arbitrary
 import org.scalacheck.Test
 import org.scalacheck.effect.PropF
 import org.typelevel.otel4s.sdk.common.InstrumentationScope
@@ -35,6 +34,7 @@ import org.typelevel.otel4s.sdk.trace.data.LinkData
 import org.typelevel.otel4s.sdk.trace.data.SpanData
 import org.typelevel.otel4s.sdk.trace.data.StatusData
 import org.typelevel.otel4s.sdk.trace.processor.SpanProcessor
+import org.typelevel.otel4s.sdk.trace.scalacheck.Arbitraries._
 import org.typelevel.otel4s.trace.SpanContext
 import org.typelevel.otel4s.trace.SpanKind
 import org.typelevel.otel4s.trace.Status
@@ -42,24 +42,6 @@ import org.typelevel.otel4s.trace.Status
 import scala.concurrent.duration._
 
 class SdkSpanBackendSuite extends CatsEffectSuite with ScalaCheckEffectSuite {
-
-  private implicit val attributesArbitrary: Arbitrary[Attributes] =
-    Arbitrary(Gens.attributes)
-
-  private implicit val scopeArbitrary: Arbitrary[InstrumentationScope] =
-    Arbitrary(Gens.instrumentationScope)
-
-  private implicit val linkDataArbitrary: Arbitrary[LinkData] =
-    Arbitrary(Gens.linkData)
-
-  private implicit val statusArbitrary: Arbitrary[Status] =
-    Arbitrary(Gens.status)
-
-  private implicit val kindArbitrary: Arbitrary[SpanKind] =
-    Arbitrary(Gens.spanKind)
-
-  private implicit val spanContextArbitrary: Arbitrary[SpanContext] =
-    Arbitrary(Gens.spanContext)
 
   // Span.Backend methods
