+++ sdk/trace/src/test/scala/org/typelevel/otel4s/sdk/trace/data/StatusDataSuite.scala
@@ -20,16 +20,16 @@ package data
 import cats.Show
 import cats.kernel.laws.discipline.HashTests
 import munit.DisciplineSuite
-import org.scalacheck.Arbitrary
 import org.scalacheck.Gen
 import org.scalacheck.Prop
+import org.typelevel.otel4s.sdk.trace.scalacheck.Arbitraries
+import org.typelevel.otel4s.sdk.trace.scalacheck.Cogens
+import org.typelevel.otel4s.sdk.trace.scalacheck.Gens
 import org.typelevel.otel4s.trace.Status
 
 class StatusDataSuite extends DisciplineSuite {
   import Cogens.statusDataCogen
-
-  private implicit val statusDataArbitrary: Arbitrary[StatusData] =
-    Arbitrary(Gens.statusData)
+  import Arbitraries.statusDataArbitrary
 
   checkAll("StatusData.HashLaws", HashTests[StatusData].hash)
