+++ sdk/trace/src/test/scala/org/typelevel/otel4s/sdk/trace/processor/BatchSpanProcessorSuite.scala
@@ -23,19 +23,16 @@ import cats.syntax.foldable._
 import cats.syntax.traverse._
 import munit.CatsEffectSuite
 import munit.ScalaCheckEffectSuite
-import org.scalacheck.Arbitrary
 import org.scalacheck.Test
 import org.scalacheck.effect.PropF
 import org.typelevel.otel4s.sdk.trace.data.SpanData
 import org.typelevel.otel4s.sdk.trace.exporter.InMemorySpanExporter
 import org.typelevel.otel4s.sdk.trace.exporter.SpanExporter
+import org.typelevel.otel4s.sdk.trace.scalacheck.Arbitraries._
 
 class BatchSpanProcessorSuite
     extends CatsEffectSuite
     with ScalaCheckEffectSuite {
-
-  private implicit val spanDataArbitrary: Arbitrary[SpanData] =
-    Arbitrary(Gens.spanData)
 
   test("show details in the name") {
     val exporter = new FailingExporter(
