+++ sdk/trace/src/test/scala/org/typelevel/otel4s/sdk/trace/processor/SimpleSpanProcessorSuite.scala
@@ -23,7 +23,6 @@ import cats.syntax.foldable._
 import cats.syntax.traverse._
 import munit.CatsEffectSuite
 import munit.ScalaCheckEffectSuite
-import org.scalacheck.Arbitrary
 import org.scalacheck.Test
 import org.scalacheck.effect.PropF
 import org.typelevel.otel4s.Attribute
@@ -33,6 +32,7 @@ import org.typelevel.otel4s.sdk.common.InstrumentationScope
 import org.typelevel.otel4s.sdk.trace.data.SpanData
 import org.typelevel.otel4s.sdk.trace.exporter.InMemorySpanExporter
 import org.typelevel.otel4s.sdk.trace.exporter.SpanExporter
+import org.typelevel.otel4s.sdk.trace.scalacheck.Arbitraries._
 import org.typelevel.otel4s.trace.Span
 import org.typelevel.otel4s.trace.SpanContext
 import org.typelevel.otel4s.trace.SpanKind
@@ -43,9 +43,6 @@ import scala.concurrent.duration.FiniteDuration
 class SimpleSpanProcessorSuite
     extends CatsEffectSuite
     with ScalaCheckEffectSuite {
-
-  private implicit val spanDataArbitrary: Arbitrary[SpanData] =
-    Arbitrary(Gens.spanData)
 
   test("show details in the name") {
     val exporter = new FailingExporter(
